backend/requirements.txt
flask==2.3.3
gremlinpython==3.7.1
kubernetes==28.1.0
requests==2.31.0
python-dotenv==1.0.0
flask-cors==4.0.0
networkx==3.2.1
pydantic==2.5.2